+++ code/game/turfs/walls/walls.dm
@@ -89,11 +89,13 @@
 				var/obj/structure/sign/poster/found_poster = found_object
 				found_poster.roll_and_drop(src)
 			if(istype(found_object, /obj/effect/alien/weeds))
-				var/list/turf/cardinal_neighbors = list(get_step(found_object, NORTH), get_step(found_object, SOUTH), get_step(found_object, EAST), get_step(found_object, WEST))
-				for(var/turf/cardinal_turf in cardinal_neighbors)
-					for(var/obj/structure/bed/nest/found_nest in cardinal_turf)
-						qdel(found_nest) //nests are built on walls, no walls, no nest
 				qdel(found_object)
+
+		var/list/turf/cardinal_neighbors = list(get_step(src, NORTH), get_step(src, SOUTH), get_step(src, EAST), get_step(src, WEST))
+		for(var/turf/cardinal_turf in cardinal_neighbors)
+			for(var/obj/structure/bed/nest/found_nest in cardinal_turf)
+				if(found_nest.dir == get_dir(found_nest, src))
+					qdel(found_nest) //nests are built on walls, no walls, no nest
 
 /turf/closed/wall/MouseDrop_T(mob/current_mob, mob/user)
 	if(acided_hole)
